Move fps/session/experiment config into cpp/h
