+++ source/FPSciApp.h
@@ -11,7 +11,7 @@
 #pragma once
 #include <G3D/G3D.h>
 
-#include "ConfigFiles.h"
+#include "ExperimentConfig.h"
 #include "StartupConfig.h"
 #include "KeyMapping.h"
 #include "SystemConfig.h"
