+++ vs/FPSci.lib.vcxproj
@@ -145,9 +145,10 @@
   <ItemGroup>
     <ClInclude Include="..\source\FPSciApp.h" />
     <ClInclude Include="..\source\CombatText.h" />
+    <ClInclude Include="..\source\FpsConfig.h" />
     <ClInclude Include="..\source\KeyMapping.h" />
     <ClInclude Include="..\source\Session.h" />
-    <ClInclude Include="..\source\ConfigFiles.h" />
+    <ClInclude Include="..\source\ExperimentConfig.h" />
     <ClInclude Include="..\source\Logger.h" />
     <ClInclude Include="..\source\PhysicsScene.h" />
     <ClInclude Include="..\source\PlayerEntity.h" />
@@ -165,7 +166,9 @@
     <ClInclude Include="..\source\Weapon.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="..\source\ExperimentConfig.cpp" />
     <ClCompile Include="..\source\FPSciApp.cpp" />
+    <ClCompile Include="..\source\FpsConfig.cpp" />
     <ClCompile Include="..\source\KeyMapping.cpp" />
     <ClCompile Include="..\source\Session.cpp" />
     <ClCompile Include="..\source\Logger.cpp" />
